Move ExecutionContext and related classes to runners-core
